Match Auton

+++ include/Autonomous/autonomous.h
@@ -1,7 +1,7 @@
 #include "vex.h"
 
 extern int auton_path;
-extern int startingPositions[3][3];
+extern double startingPositions[3][3];
 
 void autonomous(void);
 void changeAuton(int);
